// src/modules/products/enums/product-category.enum.ts
export enum ProductCategory {
  // VERDULERÍA
  FRUIT = "FRUIT",
  VEGETABLE = "VEGETABLE",
  HERBS = "HERBS",
  TUBER = "TUBER",
  SEEDLING = "SEEDLING",
  // PANADERÍA
  BREAD = "BREAD",
  PASTRIES = "PASTRIES",
  CAKES = "CAKES",
  COOKIES = "COOKIES",
  DOUGH = "DOUGH",
  DESSERTS_BAKERY = "DESSERTS_BAKERY",
  // RESTAURANTE
  APPETIZERS = "APPETIZERS",
  MAIN_COURSES = "MAIN_COURSES",
  DESSERTS_RESTAURANT = "DESSERTS_RESTAURANT",
  BEVERAGES_RESTAURANT = "BEVERAGES_RESTAURANT",
  // KIOSCO/SUPERMERCADO
  CLEANING = "CLEANING",
  BEVERAGES = "BEVERAGES",
  GROCERIES = "GROCERIES",
  FRESH_PRODUCTS = "FRESH_PRODUCTS",
  FROZEN = "FROZEN",
  PERSONAL_HYGIENE = "PERSONAL_HYGIENE",
  CANDY = "CANDY",
  SNACKS = "SNACKS",
  CIGARETTES = "CIGARETTES",
  MAGAZINES = "MAGAZINES",
  // CROSS
  GREENGROCERY_SECTION = "GREENGROCERY_SECTION",
  BAKERY_SECTION = "BAKERY_SECTION",
  // OTHER
  OTHER = "OTHER",
}